@@ -1,56 +1,119 @@
 name: ecash
 description: A new Flutter project.
 
+# The following line prevents the package from being accidentally published to
+# pub.dev using `flutter pub publish`. This is preferred for private packages.
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
+# The following defines the version and build number for your application.
+# A version number is three numbers separated by dots, like 1.2.43
+# followed by an optional build number separated by a +.
+# Both the version and the builder number may be overridden in flutter
+# build by specifying --build-name and --build-number, respectively.
+# In Android, build-name is used as versionName while build-number used as versionCode.
+# Read more about Android versioning at https://developer.android.com/studio/publish/versioning
+# In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
+# Read more about iOS versioning at
+# https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.5 <3.0.0"
 
+# Dependencies specify other packages that your package needs in order to work.
+# To automatically upgrade your package dependencies to the latest versions
+# consider running `flutter pub upgrade --major-versions`. Alternatively,
+# dependencies can be manually updated by changing the version numbers below to
+# the latest version available on pub.dev. To see which dependencies have newer
+# versions available, run `flutter pub outdated`.
 dependencies:
   flutter:
     sdk: flutter
-  flutter_riverpod: ^0.12.4
-  # hooks_riverpod: ^0.12.4
-  carousel_slider: ^3.0.0
-  cached_network_image: ^2.5.1
-  shimmer: ^1.1.2
-  page_transition: ^1.1.7+6
-  qr_flutter: ^3.2.0
-  qr_code_scanner: ^0.3.5
-  intl: ^0.16.1
-  mask_text_input_formatter: ^2.0.0
+  google_fonts: ^3.0.1
+  auto_route: ^5.0.2
+  json_annotation: ^4.7.0
+  freezed_annotation: ^2.2.0
+  flutter_riverpod: ^2.1.1
+  hooks_riverpod: ^2.1.1
+  carousel_slider: ^4.1.1
+  cached_network_image: ^3.2.1
+  shimmer: ^2.0.0
+  qr_flutter: ^4.0.0
+  qr_code_scanner: ^1.0.1
+  intl: ^0.17.0
+  mask_text_input_formatter: ^2.4.0
+  toast: ^0.3.0
+  firebase_auth: ^4.1.2
+  cloud_firestore: ^4.0.5
+  firebase_core: ^2.2.0
+  dio: ^4.0.6
+  shared_preferences: ^2.0.15
+
+
+  # The following adds the Cupertino Icons font to your application.
+  # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  firebase_core : ^0.5.0
-  cloud_firestore: ^0.14.0+2
-  firebase_auth: ^0.18.0+1
-  toast: ^0.1.5
+
+flutter_icons:
+  android: "launcher_icon"
+  ios: true
+  image_path: "assets/ecash_app_icon.png"
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_launcher_icons: "^0.8.0"
+  flutter_launcher_icons: ^0.10.0
+  auto_route_generator: ^5.0.3
+  build_runner: ^2.3.2
+  json_serializable: ^6.5.4
+  freezed: ^2.2.1
+  # The "flutter_lints" package below contains a set of recommended lints to
+  # encourage good coding practices. The lint set provided by the package is
+  # activated in the `analysis_options.yaml` file located at the root of your
+  # package. See that file for information about deactivating specific lint
+  # rules and activating additional ones.
+  flutter_lints: ^2.0.0
 
-flutter_icons:
-  android: "launcher_icon"
-  ios: false
-  image_path: "assets/ecash_app_icon.png"
+# For information on the generic Dart part of this file, see the
+# following page: https://dart.dev/tools/pub/pubspec
 
-
+# The following section is specific to Flutter packages.
 flutter:
 
+  # The following line ensures that the Material Icons font is
+  # included with your application, so that you can use the icons in
+  # the material Icons class.
   uses-material-design: true
 
   assets:
     - assets/
+  # To add assets to your application, add an assets section, like this:
+  # assets:
+  #   - images/a_dot_burr.jpeg
+  #   - images/a_dot_ham.jpeg
 
-  fonts:
-    - family: Cabin-Bold
-      fonts:
-        - asset: fonts/Cabin-Bold.ttf
-    - family: Cabin-Regular
-      fonts:
-        - asset: fonts/Cabin-Regular.ttf
-    - family: Cabin-SemiBold
-      fonts:
-        - asset: fonts/Cabin-SemiBold.ttf
+  # An image asset can refer to one or more resolution-specific "variants", see
+  # https://flutter.dev/assets-and-images/#resolution-aware
+
+  # For details regarding adding assets from package dependencies, see
+  # https://flutter.dev/assets-and-images/#from-packages
+
+  # To add custom fonts to your application, add a fonts section here,
+  # in this "flutter" section. Each entry in this list should have a
+  # "family" key with the font family name, and a "fonts" key with a
+  # list giving the asset and other descriptors for the font. For
+  # example:
+  # fonts:
+  #   - family: Schyler
+  #     fonts:
+  #       - asset: fonts/Schyler-Regular.ttf
+  #       - asset: fonts/Schyler-Italic.ttf
+  #         style: italic
+  #   - family: Trajan Pro
+  #     fonts:
+  #       - asset: fonts/TrajanPro.ttf
+  #       - asset: fonts/TrajanPro_Bold.ttf
+  #         weight: 700
+  #
+  # For details regarding fonts from package dependencies,
+  # see https://flutter.dev/custom-fonts/#from-packages
