@@ -1,11 +1,10 @@
 import 'package:ecash/constants/app_color.dart';
-import 'package:ecash/constants/app_font.dart';
 import 'package:flutter/material.dart';
 
 class PrimaryButton extends StatelessWidget {
-  const PrimaryButton({Key key, this.onPressed, this.title}) : super(key: key);
-  final VoidCallback onPressed;
-  final String title;
+  const PrimaryButton({Key? key, this.onPressed, this.title}) : super(key: key);
+  final VoidCallback? onPressed;
+  final String? title;
   @override
   Widget build(BuildContext context) {
     return SizedBox(
@@ -19,7 +18,7 @@ class PrimaryButton extends StatelessWidget {
         onPressed: onPressed,
         child: Text(
           title ?? '',
-          style: AppFont.bold(
+          style: TextStyle(
             color: Colors.white,
             fontSize: 20,
           ),
