@@ -1,11 +1,10 @@
 import 'package:ecash/components/primary_icon_button.dart';
 import 'package:ecash/constants/app_color.dart';
-import 'package:ecash/constants/app_font.dart';
 import 'package:flutter/material.dart';
 
 class PrimaryDrawer extends StatelessWidget {
-  const PrimaryDrawer({Key key, this.onLogout}) : super(key: key);
-  final VoidCallback onLogout;
+  const PrimaryDrawer({Key? key, this.onLogout}) : super(key: key);
+  final VoidCallback? onLogout;
   @override
   Widget build(BuildContext context) {
     return Drawer(
@@ -22,7 +21,7 @@ class PrimaryDrawer extends StatelessWidget {
               alignment: Alignment.centerRight,
               child: GestureDetector(
                 onTap: () => Navigator.pop(context),
-                child: Icon(
+                child: const Icon(
                   Icons.close,
                   color: Colors.white,
                 ),
@@ -43,17 +42,17 @@ class PrimaryDrawer extends StatelessWidget {
                 ),
                 Column(
                   crossAxisAlignment: CrossAxisAlignment.end,
-                  children: [
+                  children: const [
                     Text(
                       'Ecash',
-                      style: AppFont.bold(
+                      style: TextStyle(
                         color: Colors.white,
                         fontSize: 30,
                       ),
                     ),
                     Text(
                       'v 1.0.0',
-                      style: AppFont.regular(
+                      style: TextStyle(
                         color: Colors.white,
                         fontSize: 12,
                       ),
@@ -79,15 +78,15 @@ class PrimaryDrawer extends StatelessWidget {
                       color: Colors.white.withOpacity(0.4),
                       borderRadius: BorderRadius.circular(10),
                     ),
-                    child: Text(
+                    child: const Text(
                       'About Us',
-                      style: AppFont.regular(
+                      style: TextStyle(
                         fontSize: 15,
                         color: Colors.white,
                       ),
                     ),
                   ),
-                  Align(
+                  const Align(
                     alignment: Alignment.centerLeft,
                     child: SizedBox(
                       width: 50,
@@ -100,29 +99,29 @@ class PrimaryDrawer extends StatelessWidget {
                 ],
               ),
             ),
-            Container(
+            SizedBox(
               height: 60,
               child: Stack(
                 alignment: Alignment.center,
                 children: [
                   Container(
                     alignment: Alignment.center,
-                    margin: EdgeInsets.only(left: 20),
+                    margin: const EdgeInsets.only(left: 20),
                     height: 30,
                     width: double.infinity,
                     decoration: BoxDecoration(
                       color: Colors.white.withOpacity(0.4),
                       borderRadius: BorderRadius.circular(10),
                     ),
-                    child: Text(
+                    child: const Text(
                       'Service',
-                      style: AppFont.regular(
+                      style: TextStyle(
                         fontSize: 15,
                         color: Colors.white,
                       ),
                     ),
                   ),
-                  Align(
+                  const Align(
                     alignment: Alignment.centerLeft,
                     child: SizedBox(
                       width: 50,
@@ -135,7 +134,7 @@ class PrimaryDrawer extends StatelessWidget {
                 ],
               ),
             ),
-            Container(
+            SizedBox(
               height: 60,
               child: Stack(
                 alignment: Alignment.center,
@@ -149,15 +148,15 @@ class PrimaryDrawer extends StatelessWidget {
                       color: Colors.white.withOpacity(0.4),
                       borderRadius: BorderRadius.circular(10),
                     ),
-                    child: Text(
+                    child: const Text(
                       'FAQ\'s',
-                      style: AppFont.regular(
+                      style: TextStyle(
                         fontSize: 15,
                         color: Colors.white,
                       ),
                     ),
                   ),
-                  Align(
+                  const Align(
                     alignment: Alignment.centerLeft,
                     child: SizedBox(
                       width: 50,
@@ -172,29 +171,29 @@ class PrimaryDrawer extends StatelessWidget {
             ),
             GestureDetector(
               onTap: onLogout,
-              child: Container(
+              child: SizedBox(
                 height: 60,
                 child: Stack(
                   alignment: Alignment.center,
                   children: [
                     Container(
                       alignment: Alignment.center,
-                      margin: EdgeInsets.only(left: 20),
+                      margin: const EdgeInsets.only(left: 20),
                       height: 30,
                       width: double.infinity,
                       decoration: BoxDecoration(
                         color: Colors.white.withOpacity(0.4),
                         borderRadius: BorderRadius.circular(10),
                       ),
-                      child: Text(
+                      child: const Text(
                         'Logout',
-                        style: AppFont.regular(
+                        style: TextStyle(
                           fontSize: 15,
                           color: Colors.white,
                         ),
                       ),
                     ),
-                    Align(
+                    const Align(
                       alignment: Alignment.centerLeft,
                       child: SizedBox(
                         width: 50,
