@@ -1,22 +1,13 @@
-import 'package:ecash/components/loading_screen.dart';
+
 import 'package:ecash/components/primary_appbar.dart';
 import 'package:ecash/components/primary_button.dart';
-import 'package:ecash/components/primary_buttonlabeled.dart';
-import 'package:ecash/components/primary_icon_button.dart';
 import 'package:ecash/constants/app_color.dart';
-import 'package:ecash/constants/app_font.dart';
-import 'package:ecash/constants/enums.dart';
-import 'package:ecash/models/transaction_model.dart';
 import 'package:ecash/pages/primary_textfield.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:intl/intl.dart';
-import 'package:toast/toast.dart';
 
-import '../main.dart';
 
 class HomeCreditBillPage extends StatelessWidget {
-  HomeCreditBillPage({Key key}) : super(key: key);
+  HomeCreditBillPage({Key? key}) : super(key: key);
   final TextEditingController amountController = TextEditingController();
   @override
   Widget build(BuildContext context) {
@@ -46,7 +37,7 @@ class HomeCreditBillPage extends StatelessWidget {
                                 alignment: Alignment.center,
                                 child: Text(
                                   'You are paying',
-                                  style: AppFont.semiBold(fontSize: 15, color: Colors.grey.shade500),
+                                  style: TextStyle(fontSize: 15, color: Colors.grey.shade500),
                                 ),
                               ),
                               const SizedBox(
@@ -79,7 +70,7 @@ class HomeCreditBillPage extends StatelessWidget {
                                   width: MediaQuery.of(context).size.width * 0.75,
                                   child: Text(
                                     'Biller Conveneince Fee (BCF) may apply',
-                                    style: AppFont.regular(
+                                    style: TextStyle(
                                       fontSize: 14,
                                     ),
                                   ),
@@ -93,7 +84,7 @@ class HomeCreditBillPage extends StatelessWidget {
                                 children: [
                                   Text(
                                     '+63',
-                                    style: AppFont.semiBold(
+                                    style: TextStyle(
                                       fontSize: 15,
                                     ),
                                   ),
@@ -128,7 +119,7 @@ class HomeCreditBillPage extends StatelessWidget {
                                       ),
                                       Text(
                                         'Remider',
-                                        style: AppFont.semiBold(
+                                        style: TextStyle(
                                           fontSize: 15,
                                         ),
                                       ),
@@ -141,7 +132,7 @@ class HomeCreditBillPage extends StatelessWidget {
                                     width: MediaQuery.of(context).size.width,
                                     child: Text(
                                       'Sed ut perspiciatis unde omnis iste natus error sit voluptatem accusantium doloremque laudantium, totam rem aperiam, eaque ipsa quae ab illo inventore veritatis et quasi architecto beatae vitae dicta ',
-                                      style: AppFont.regular(
+                                      style: TextStyle(
                                         fontSize: 12,
                                       ),
                                     ),
@@ -154,24 +145,24 @@ class HomeCreditBillPage extends StatelessWidget {
                               PrimaryButton(
                                 title: 'Pay',
                                 onPressed: () {
-                                  final userWallet = context.read(userProvider).user.wallet;
-                                  if (userWallet < double.parse(amountController.text)) {
-                                    Toast.show('Sorry, Unsufficient Balance', context, duration: 3, gravity: Toast.TOP);
-                                  } else {
-                                    TransactionModel newTransaction = TransactionModel(
-                                      amount: double.parse(amountController.text),
-                                      date: DateTime.now(),
-                                      description: 'You paid Home Credit bill with amount of PHP ' + amountController.text,
-                                      title: 'Pay Bill',
-                                      type: TransactionType.expense,
-                                    );
-                                    context.read(userTransactions).addTransaction(
-                                          context: context,
-                                          currentUserWallet: userWallet,
-                                          transaction: newTransaction,
-                                          type: TransactionType.expense,
-                                        );
-                                  }
+                                  // final userWallet = context.read(userProvider).user.wallet;
+                                  // if (userWallet < double.parse(amountController.text)) {
+                                  //   Toast.show('Sorry, Unsufficient Balance', context, duration: 3, gravity: Toast.TOP);
+                                  // } else {
+                                  //   TransactionModel newTransaction = TransactionModel(
+                                  //     amount: double.parse(amountController.text),
+                                  //     date: DateTime.now(),
+                                  //     description: 'You paid Home Credit bill with amount of PHP ' + amountController.text,
+                                  //     title: 'Pay Bill',
+                                  //     type: TransactionType.expense,
+                                  //   );
+                                  //   context.read(userTransactions).addTransaction(
+                                  //         context: context,
+                                  //         currentUserWallet: userWallet,
+                                  //         transaction: newTransaction,
+                                  //         type: TransactionType.expense,
+                                  //       );
+                                  // }
                                 },
                               ),
                               const SizedBox(
@@ -191,15 +182,15 @@ class HomeCreditBillPage extends StatelessWidget {
               ),
             ),
           ),
-          Consumer(
-            builder: (context, watch, child) {
-              final isloading = context.read(userTransactions).isLoading;
-              if (isloading) {
-                return LoadingScreen();
-              }
-              return const SizedBox();
-            },
-          )
+          // Consumer(
+          //   builder: (context, watch, child) {
+          //     final isloading = context.read(userTransactions).isLoading;
+          //     if (isloading) {
+          //       return LoadingScreen();
+          //     }
+          //     return const SizedBox();
+          //   },
+          // )
         ],
       ),
     );
