@@ -1,13 +1,12 @@
 import 'package:ecash/pages/home_page.dart';
 import 'package:ecash/pages/profile_page.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class MainPage extends StatefulWidget {
-  MainPage({Key key}) : super(key: key);
+  const MainPage({Key? key}) : super(key: key);
 
   @override
-  _MainPageState createState() => _MainPageState();
+  State<MainPage> createState() => _MainPageState();
 }
 
 class _MainPageState extends State<MainPage> {
@@ -17,14 +16,14 @@ class _MainPageState extends State<MainPage> {
   Widget build(BuildContext context) {
     return Scaffold(
       body: PageView(
-        physics: NeverScrollableScrollPhysics(),
+        physics: const NeverScrollableScrollPhysics(),
         controller: _pageController,
         children: [
           HomePage(
-            onProfile: () => _pageController.animateToPage(1, duration: Duration(milliseconds: 300), curve: Curves.ease),
+            onProfile: () => _pageController.animateToPage(1, duration: const Duration(milliseconds: 300), curve: Curves.ease),
           ),
           ProfilePage(
-            onHome: () => _pageController.animateToPage(0, duration: Duration(milliseconds: 300), curve: Curves.ease),
+            onHome: () => _pageController.animateToPage(0, duration: const Duration(milliseconds: 300), curve: Curves.ease),
           ),
         ],
       ),
