@@ -1,16 +1,13 @@
 import 'dart:developer';
 import 'dart:io';
 import 'package:ecash/constants/app_color.dart';
-import 'package:ecash/constants/app_font.dart';
 import 'package:ecash/pages/qr_payment_result_page.dart';
-import 'package:ecash/utils/navigaton_service.dart';
 import 'package:flutter/material.dart';
-import 'package:page_transition/page_transition.dart';
 import 'package:qr_code_scanner/qr_code_scanner.dart';
 import 'package:toast/toast.dart';
 
 class PayQRScanner extends StatefulWidget {
-  PayQRScanner({Key key}) : super(key: key);
+  const PayQRScanner({Key? key}) : super(key: key);
 
   @override
   State<PayQRScanner> createState() => _PayQRScannerState();
@@ -19,27 +16,27 @@ class PayQRScanner extends StatefulWidget {
 class _PayQRScannerState extends State<PayQRScanner> {
   final GlobalKey qrKey = GlobalKey(debugLabel: 'QR');
 
-  Barcode result;
+  Barcode? result;
 
-  QRViewController controller;
+  QRViewController? controller;
 
   void _onQRViewCreated(QRViewController controller) {
     this.controller = controller;
     controller.scannedDataStream.listen((scanData) {
-      log('qr Scanned: ' + scanData.code);
-      if (scanData.code.contains("ecash-pay")) {
+      log('qr Scanned: ' + scanData.code!);
+      if (scanData.code!.contains("ecash-pay")) {
         try {
-          NavigationService().replaceScreen(
-            page: QrPaymentResult(
-              amount: double.parse(scanData.code.split('-').last),
-            ),
-            type: PageTransitionType.rightToLeft,
-          );
+          // NavigationService().replaceScreen(
+          //   page: QrPaymentResult(
+          //     amount: double.parse(scanData.code.split('-').last),
+          //   ),
+          //   type: PageTransitionType.rightToLeft,
+          // );
         } catch (e) {
-          Toast.show('Invalid Barcode', context);
+          Toast.show('Invalid Barcode');
         }
       } else {
-        Toast.show('Invalid Barcode', context);
+        Toast.show('Invalid Barcode');
       }
     });
   }
@@ -48,9 +45,9 @@ class _PayQRScannerState extends State<PayQRScanner> {
   void reassemble() {
     super.reassemble();
     if (Platform.isAndroid) {
-      controller.pauseCamera();
+      controller!.pauseCamera();
     } else if (Platform.isIOS) {
-      controller.resumeCamera();
+      controller!.resumeCamera();
     }
   }
 
@@ -96,10 +93,10 @@ class _PayQRScannerState extends State<PayQRScanner> {
                           width: 10,
                         ),
                         GestureDetector(
-                          onTap: () => Navigator.push(context, PageTransition(child: QrPaymentResult(), type: PageTransitionType.rightToLeft)),
+                          //onTap: () => Navigator.push(context, PageTransition(child: QrPaymentResult(), type: PageTransitionType.rightToLeft)),
                           child: Text(
                             'Scan Payment Bills',
-                            style: AppFont.bold(
+                            style: TextStyle(
                               fontSize: 20,
                             ),
                           ),
@@ -113,7 +110,7 @@ class _PayQRScannerState extends State<PayQRScanner> {
                       width: MediaQuery.of(context).size.width * 0.8,
                       child: Text(
                         'Lorem Ipsum has been the industry able content of a page when looking at its layout. The point of using Lorem',
-                        style: AppFont.regular(
+                        style: TextStyle(
                           fontSize: 14,
                         ),
                       ),
