@@ -1,14 +1,13 @@
 import 'package:ecash/constants/app_color.dart';
-import 'package:ecash/constants/app_font.dart';
 import 'package:flutter/material.dart';
 
 class PrimaryTextField extends StatelessWidget {
-  const PrimaryTextField({Key key, this.width, this.hint, this.controller, this.maxLength, this.textInputType}) : super(key: key);
-  final double width;
-  final String hint;
-  final TextEditingController controller;
-  final int maxLength;
-  final TextInputType textInputType;
+  const PrimaryTextField({Key? key, this.width, this.hint, this.controller, this.maxLength, this.textInputType}) : super(key: key);
+  final double? width;
+  final String? hint;
+  final TextEditingController? controller;
+  final int? maxLength;
+  final TextInputType? textInputType;
   @override
   Widget build(BuildContext context) {
     return Container(
@@ -27,7 +26,7 @@ class PrimaryTextField extends StatelessWidget {
           enabledBorder: InputBorder.none,
           errorBorder: InputBorder.none,
           hintText: hint,
-          hintStyle: AppFont.regular(
+          hintStyle: TextStyle(
             color: Colors.grey.shade600,
             fontSize: 14,
           ),
@@ -37,7 +36,7 @@ class PrimaryTextField extends StatelessWidget {
             bottom: 10,
           ),
         ),
-        style: AppFont.semiBold(color: AppColor.primary, fontSize: 14),
+        style: TextStyle(color: AppColor.primary, fontSize: 14),
       ),
     );
   }
