@@ -1,32 +1,31 @@
+import 'package:auto_route/auto_route.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import 'package:ecash/components/primary_buttonlabeled.dart';
 import 'package:ecash/components/primary_icon_button.dart';
 import 'package:ecash/components/transaction_card.dart';
 import 'package:ecash/components/user_profile_card.dart';
 import 'package:ecash/constants/app_color.dart';
-import 'package:ecash/constants/app_font.dart';
-import 'package:ecash/constants/enums.dart';
-import 'package:ecash/constants/functions.dart';
-import 'package:ecash/main.dart';
+import 'package:ecash/constants/photos.dart';
 import 'package:ecash/models/transaction_model.dart';
 import 'package:ecash/pages/transaction_page.dart';
+import 'package:ecash/providers/providers.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:page_transition/page_transition.dart';
 
-class ProfilePage extends StatefulWidget {
-  const ProfilePage({Key key, this.onHome}) : super(key: key);
-  final VoidCallback onHome;
+class ProfilePage extends ConsumerStatefulWidget {
+  const ProfilePage({Key? key, this.onHome}) : super(key: key);
+  final VoidCallback? onHome;
 
   @override
-  _ProfilePageState createState() => _ProfilePageState();
+  ConsumerState<ProfilePage> createState() => _ProfilePageState();
 }
 
-class _ProfilePageState extends State<ProfilePage> {
+class _ProfilePageState extends ConsumerState<ProfilePage> {
   @override
   void initState() {
     super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) => context.read(userTransactions).getTransactions());
+    WidgetsBinding.instance
+        .addPostFrameCallback((_) => ref.read(transactionProvider.notifier).getUserTransactionsHistory());
   }
 
   @override
@@ -34,23 +33,18 @@ class _ProfilePageState extends State<ProfilePage> {
     return Scaffold(
       body: Container(
         color: AppColor.background,
-        padding: EdgeInsets.only(left: 20, right: 20),
+        padding: const EdgeInsets.only(left: 20, right: 20),
         child: SingleChildScrollView(
           child: Column(
             children: [
               const SizedBox(
                 height: 50,
               ),
-              Consumer(
-                builder: (context, watch, child) {
-                  final user = context.read(userProvider).user;
-                  return UserProfileCard(
-                    displayName: user.firstName + ' ' + user.lastName,
-                    profilePhoto: user.photo,
-                    mobileNumber: user.mobile,
-                    userAccountNumber: user.accountNumber,
-                  );
-                },
+              UserProfileCard(
+                displayName: '${ref.watch(authProvider).getUserData()!.firstname} ${ref.watch(authProvider).getUserData()!.lastname!}',
+                profilePhoto: profilePlaceHolder,
+                mobileNumber: ref.watch(authProvider).getUserData()!.mobileNumber,
+                userAccountNumber: '1234-1233-999',
               ),
               const SizedBox(
                 height: 30,
@@ -62,18 +56,22 @@ class _ProfilePageState extends State<ProfilePage> {
                   PriamryButtonLabeled(
                     asset: 'assets/heart.png',
                     label: 'Favorites',
+                    onTap: (){},
                   ),
                   PriamryButtonLabeled(
                     asset: 'assets/book.png',
                     label: 'Quick \n Guide',
+                    onTap: (){},
                   ),
                   PriamryButtonLabeled(
                     asset: 'assets/account.png',
                     label: 'Account \n Limit',
+                    onTap: (){},
                   ),
                   PriamryButtonLabeled(
                     asset: 'assets/settings.png',
                     label: 'Settings',
+                    onTap: (){},
                   ),
                 ],
               ),
@@ -84,9 +82,9 @@ class _ProfilePageState extends State<ProfilePage> {
                 alignment: Alignment.bottomLeft,
                 child: SizedBox(
                   width: MediaQuery.of(context).size.width * 0.75,
-                  child: Text(
+                  child: const Text(
                     'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore ',
-                    style: AppFont.regular(fontSize: 14),
+                    style: TextStyle(fontSize: 14),
                   ),
                 ),
               ),
@@ -101,7 +99,7 @@ class _ProfilePageState extends State<ProfilePage> {
                   ),
                   borderRadius: BorderRadius.circular(20),
                 ),
-                padding: EdgeInsets.all(10),
+                padding: const EdgeInsets.all(10),
                 child: Column(
                   children: [
                     Padding(
@@ -109,18 +107,18 @@ class _ProfilePageState extends State<ProfilePage> {
                       child: Row(
                         mainAxisAlignment: MainAxisAlignment.spaceBetween,
                         children: [
-                          Text(
+                          const Text(
                             'Transaction',
-                            style: AppFont.bold(
+                            style: TextStyle(
                               fontSize: 20,
                               color: Colors.black,
                             ),
                           ),
                           GestureDetector(
-                            onTap: () => Navigator.push(context, PageTransition(child: TransactionPage(), type: PageTransitionType.rightToLeft)),
+                            onTap: () => context.router.pushNamed('/transactions'),
                             child: Text(
                               'see all',
-                              style: AppFont.regular(
+                              style: TextStyle(
                                 fontSize: 12,
                                 color: AppColor.primary,
                               ),
@@ -132,47 +130,53 @@ class _ProfilePageState extends State<ProfilePage> {
                     const SizedBox(
                       height: 10,
                     ),
-                    Consumer(
-                      builder: (context, watch, child) {
-                        final transaction = watch(userTransactions);
-                        if (transaction.isLoading) {
-                          return Center(
-                            child: Padding(
-                              padding: EdgeInsets.only(top: 20, bottom: 20),
-                              child: CircularProgressIndicator(
-                                backgroundColor: Colors.transparent,
-                                valueColor: AlwaysStoppedAnimation<Color>(AppColor.primary),
-                              ),
+                    ref.watch(transactionProvider).transactions.when(
+                      loading: () {
+                        return Center(
+                          child: Padding(
+                            padding: const EdgeInsets.only(top: 20, bottom: 20),
+                            child: CircularProgressIndicator(
+                              backgroundColor: Colors.transparent,
+                              valueColor: AlwaysStoppedAnimation<Color>(AppColor.primary),
                             ),
-                          );
-                        }
-                        if (transaction.listOfTransaction.isNotEmpty) {
+                          ),
+                        ); 
+                      }, 
+                      error: (_, stackTrace) {
+                        return const Center(
+                          child: Text(
+                            'Went somethinf wrong',
+                          ),
+                        );
+                      },
+                      data: (transaction) {
+                        if (transaction.isNotEmpty) {
                           List<TransactionModel> recentList = <TransactionModel>[];
                           for (int count = 0; count < 3; count++) {
                             try {
-                              recentList.add(transaction.listOfTransaction[count]);
+                              recentList.add(transaction[count]);
                             } catch (e) {}
                           }
                           return Column(
                             children: List.generate(recentList.length, (index) {
                               return TransactionCard(
                                 amount: recentList[index].amount,
-                                date: recentList[index].date,
+                                date: recentList[index].date!,
                                 description: recentList[index].description,
                                 title: recentList[index].title,
                                 type: recentList[index].type,
                               );
                             }),
                           );
+                        } else {
+                          return const Center(
+                            child: Text(
+                              'No transaction yet.'
+                            ),
+                          );
                         }
-                        return Center(
-                          child: Text(
-                            'No Transaction Yet.',
-                            style: AppFont.bold(fontSize: 18, color: Colors.grey),
-                          ),
-                        );
-                      },
-                    )
+                      }
+                    ),
                   ],
                 ),
               ),
@@ -191,14 +195,14 @@ class _ProfilePageState extends State<ProfilePage> {
                         )),
                     onPressed: widget.onHome,
                     child: Row(
-                      children: [
-                        Icon(
+                      children: const [
+                       Icon(
                           Icons.arrow_back_ios_rounded,
                           color: Colors.white,
                         ),
                         Text(
                           'Back to Home',
-                          style: AppFont.bold(
+                          style: TextStyle(
                             color: Colors.white,
                             fontSize: 15,
                           ),
