@@ -1,27 +1,33 @@
-import 'package:ecash/pages/root_page.dart';
-import 'package:ecash/providers/transaction_provider.dart';
-import 'package:ecash/providers/user_provider.dart';
-import 'package:ecash/utils/navigaton_service.dart';
+
+import 'package:ecash/router/app_router.gr.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:google_fonts/google_fonts.dart';
 
-final userProvider = ChangeNotifierProvider((ref) => UserProvider());
-final userTransactions = ChangeNotifierProvider((ref) => TransactionProvider());
+final appRouter = AppRouter();
+
+
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  await SystemChrome.setPreferredOrientations([
+    DeviceOrientation.portraitUp,
+  ]);
   await Firebase.initializeApp();
-
-  SystemChrome.setPreferredOrientations([DeviceOrientation.portraitUp]).then((_) {
-    runApp(
-      ProviderScope(
-        child: MaterialApp(
-          navigatorKey: NavigationService().navigationKey,
-          debugShowCheckedModeBanner: false,
-          home: RootPage(),
+  runApp(
+    ProviderScope(
+      child: MaterialApp.router(  
+        routeInformationParser: appRouter.defaultRouteParser(),
+        debugShowCheckedModeBanner: false,
+        theme: ThemeData(
+          fontFamily: GoogleFonts.nunito().fontFamily,
+          textTheme: GoogleFonts.nunitoTextTheme()
         ),
+        routerDelegate: appRouter.delegate(),
+        builder:(context,child) => child!,
       ),
-    );
-  });
+    )
+  );
 }
+
